Rename valudation to _schema
