+++ src/_schema/__tests__/json-schema-draft4/minProperties.json
